Estimation Engine unit test: Separate results and datasets
